+++ cli/cmd/node/getContainerRuntime.go
@@ -6,8 +6,8 @@ import (
 	"github.com/opctl/opctl/cli/internal/nodeprovider/local"
 	"github.com/opctl/opctl/sdks/go/node/containerruntime"
 	"github.com/opctl/opctl/sdks/go/node/containerruntime/docker"
+	"github.com/opctl/opctl/sdks/go/node/containerruntime/embedded"
 	"github.com/opctl/opctl/sdks/go/node/containerruntime/k8s"
-	"github.com/opctl/opctl/sdks/go/node/containerruntime/qemu"
 )
 
 func getContainerRuntime(
@@ -16,8 +16,8 @@ func getContainerRuntime(
 ) (containerruntime.ContainerRuntime, error) {
 	if nodeConfig.ContainerRuntime == "k8s" {
 		return k8s.New()
-	} else if nodeConfig.ContainerRuntime == "qemu" {
-		return qemu.New(ctx, false)
+	} else if nodeConfig.ContainerRuntime == "embedded" {
+		return embedded.New(ctx, nodeConfig.DataDir)
 	} else {
 		return docker.New(ctx, "unix:///var/run/docker.sock")
 	}