+++ sdks/go/node/containerruntime/embedded/limactl_darwin_amd64.go
@@ -1,0 +1,9 @@
+package embedded
+
+import _ "embed"
+
+//go:embed embeds/darwin-amd64/limactl
+var limactlBytes []byte
+
+//go:embed embeds/darwin-amd64/lima-guestagent.Linux-x86_64
+var limaLinuxGuestAgentBytes []byte